+++ src/_utils/appcache/index.js
@@ -1,0 +1,29 @@
+import globby from "globby"
+import { join } from "path"
+import { writeFile } from "fs-promise"
+import template from "./template"
+import joinUri from "../join-uri"
+
+const writeAppcache = (distPath, baseUrl, pattern) => {
+  const destination = join(distPath, "manifest.appcache")
+  const fallback = joinUri("/", baseUrl, "/")
+
+  return globby(pattern, {
+    cwd: distPath,
+    nodir: true,
+  })
+    .then(
+      (files) => files.map((file) => joinUri(
+        "/",
+        baseUrl,
+        file,
+      ))
+    )
+    .then(
+      (files) => Promise.all([
+        writeFile(destination, template(files, fallback)),
+      ])
+    )
+}
+
+export default writeAppcache