+++ src/configurator/index.js
@@ -14,6 +14,7 @@ const defaultOptions = {
   devPort: "3000",
   verbose: false,
   open: true,
+  appcache: false,
 }
 
 const defaultOptionsCLIonly = {
@@ -28,6 +29,7 @@ const fieldTypes = {
   "source": "string",
   "destination": "string",
   "assets": true, // accept object, boolean, falsy or string values
+  "appcache": true, // accept array, falsy or string values
   "CNAME": "boolean",
   "nojekyll": "boolean",
   "devHost": "string",
@@ -179,6 +181,31 @@ export default function config(pkg = {}, argv = process.argv) {
         )
       }
     }
+  }
+
+  if (typeof config.appcache === "string") {
+    config.appcache = [ config.appcache ]
+  }
+  // Default value if set true
+  else if (typeof config.appcache === "boolean" && config.appcache) {
+    config.appcache = [ "**/*.*", "!**/*.html", "index.html" ]
+  }
+  else if (config.appcache === null) {
+    config.appcache = false
+  }
+  else if (
+    !Array.isArray(config.appcache) &&
+    typeof config.appcache !== "boolean"
+  ) {
+    errors.push(
+      `You provided an '${ typeof config.appcache }' for 'appcache' option. ` +
+      `This option accepts a boolean value, a string, or an array.`
+    )
+  }
+
+  // Disable appcache for development
+  if (config.dev) {
+    config.appcache = false
   }
 
   if (errors.length > 0) {